feat: Support shadowing and rounded corner for rect draw, update surface to fit that

+++ tessera/src/renderer/drawer/pos_misc.rs
@@ -1,6 +1,6 @@
-pub fn pixel_to_ndc(pos: [u32; 2], screen_size: [u32; 2]) -> [f32; 2] {
-    let x = pos[0] as f32 / screen_size[0] as f32 * 2.0 - 1.0;
-    let y = pos[1] as f32 / screen_size[1] as f32 * 2.0 - 1.0;
+pub fn pixel_to_ndc(pos: [f32; 2], screen_size: [u32; 2]) -> [f32; 2] {
+    let x = pos[0] / screen_size[0] as f32 * 2.0 - 1.0;
+    let y = pos[1] / screen_size[1] as f32 * 2.0 - 1.0;
     // Invert y axis
     // because the origin is at the bottom left corner in OpenGL
     // but we want the origin to be at the top left corner, since
